refactor(auth): move ALCHEMY_PROFILE config + withProfileOverride helper into Auth/Profile.ts
The CLI's resolveConfigProvider (which built a tiny ConfigProvider that
shadows ALCHEMY_PROFILE with the --profile flag) was profile/auth logic
hiding in bin/. Moved to Auth/Profile.ts as `withProfileOverride`, and
the shared `ALCHEMY_PROFILE` Config (defaults to "default") now lives
there too — Cloudflare/Credentials.ts and Cloudflare/CloudflareEnvironment.ts
import the shared one instead of re-declaring it.

Made-with: Cursor
